refactor: Random CSS 및 인터렉션 변경
